+++ app/Jobs/UploadAdVideoCreative.php
@@ -68,7 +68,9 @@ class UploadAdVideoCreative implements ShouldQueue
         $adSets[$this->adSetIndex]['creatives'][$this->creativeIndex]['video_id'] = $videoId;
         $this->adCreationFlow->update(['adSets' => $adSets]);
 
-        $disk->delete($this->path);
+        if ($disk->exists($this->path)) {
+            $disk->delete($this->path);
+        }
     }
 
     public function middleware(): array
